fix: 🔨 Fix testnet Identity pallet's configuration (#127)
This pull request updates the identity pallet configuration to ensure
that slashed funds are now sent to the treasury, rather than being
discarded. This change is a step towards proper fund management within
the runtime.

Identity pallet configuration update:

* Changed the `Slashed` associated type implementation in the
`pallet_identity::Config` for `Runtime` to use `Treasury`, so that
slashed funds are now sent to the treasury instead of being ignored.